Move miniaudio files to a subdir
