+++ src/SFML/Audio/Miniaudio.cpp
@@ -29,4 +29,4 @@
 #pragma GCC diagnostic ignored "-Wnull-dereference" // Work around MinGW warning
 #endif
 // #define MA_DEBUG_OUTPUT // Uncomment to enable miniaudio debug output to console
-#include <miniaudio.c> // NOLINT(bugprone-suspicious-include)
+#include <miniaudio/miniaudio.c> // NOLINT(bugprone-suspicious-include)
